# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Grão Direto TechBlog API",
    "description": "Documentação da API de Artigos e Usuários da Grão Direto.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "tags": [
    { "name": "Users", "description": "Gerenciamento e Autenticação de Usuários." },
    { "name": "Articles", "description": "Gerenciamento e Listagem de Artigos." },
    { "name": "Comments", "description": "Interações e Comentários." }
  ],
  "paths": {
   
    "/users/create": {
      "post": {
        "tags": ["Users"],
        "summary": "Cria um novo usuário.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Novo Colaborador" },
                  "email": { "type": "string", "format": "email", "example": "user@graodireto.com" },
                  "password": { "type": "string", "format": "password", "example": "SenhaForte123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado com sucesso." },
          "400": { "description": "Dados inválidos ou email já cadastrado." }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Realiza o login e retorna um Token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido. Retorna o token." },
          "401": { "description": "Credenciais inválidas." }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": ["Users"],
        "summary": "Faz logout do usuário (Invalidação do Token/Blacklist).",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Logout realizado com sucesso." },
          "401": { "description": "Não autorizado." }
        }
      }
    },
    "/users/logado": {
      "get": {
        "tags": ["Users"],
        "summary": "Retorna os dados do usuário logado.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dados do usuário (ID, Nome, Email)." },
          "401": { "description": "Token inválido." }
        }
      }
    },
    "/users/edit/{id}": {
      "put": {
        "tags": ["Users"],
        "summary": "Atualiza os dados do usuário (inclui foto de perfil).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          { "name": "authorization", "in": "header", "schema": { "type": "string" }, "description": "Bearer Token" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "password": { "type": "string", "format": "password" },
                  "profilePictureUrl": { "type": "string", "format": "binary", "description": "Arquivo de imagem (multipart/form-data)." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Usuário atualizado com sucesso." },
          "401": { "description": "Não autorizado." },
          "404": { "description": "Usuário não encontrado." }
        }
      }
    },
  
    "/api/articles": {
      "get": {
        "tags": ["Articles"],
        "summary": "Lista todos os artigos com paginação.",
        "responses": { "200": { "description": "Retorna a lista paginada." } }
      },
      "post": {
        "tags": ["Articles"],
        "summary": "Cria um novo artigo.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "url" },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Artigo criado com sucesso." },
          "401": { "description": "Não autorizado." }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Busca detalhes de um artigo por ID.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Detalhes do artigo." }, "404": { "description": "Artigo não encontrado." } }
      },
      "put": {
        "tags": ["Articles"],
        "summary": "Atualiza um artigo existente (requer ser o autor).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "url" },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Artigo atualizado com sucesso." }, "401": { "description": "Não autorizado." }, "404": { "description": "Artigo não encontrado." } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}